+++ src/CurrencyFactory.php
@@ -10,6 +10,47 @@ class CurrencyFactory
     protected static $default = 'ZAR';
 
     /**
+     * @var array
+     *
+     * The format is 'code' => array('sign', 'name') OR 'code' => array('sign', 'name', 'isLeftSign')
+     */
+    protected static $currencies = array(
+        'AUD' => array('$', 'Australian Dollar'),
+        'BRL' => array('R$', 'Brazilian Real'),
+        'CAD' => array('$', 'Canadian Dollar'),
+        'CNY' => array('¥', 'Chinese Yuan'),
+        'CHF' => array('Fr', 'Swiss Franc'),
+        'DKK' => array('kr.', 'Danish Krone'),
+        'EUR' => array('€', 'Euro'),
+        'GBP' => array('£', 'British Pound Sterling'),
+        'HKD' => array('$', 'Hong Kong Dollar'),
+        'INR' => array('₹', 'Indian Rupee'),
+        'ILS' => array('₪', 'Israeli New Shekel'),
+        'JPY' => array('¥', 'Japanese Yen'),
+        'KRW' => array('₩', 'South Korean Won'),
+        'MXN' => array('$', 'Mexican Peso'),
+        'NZD' => array('$', 'New Zealand Dollar'),
+        'NOK' => array('kr', 'Norwegian Krone'),
+        'RUB' => array('₽', 'Russian Ruble'),
+        'SEK' => array('kr', 'Swedish Krona'),
+        'SGD' => array('$', 'Singapore Dollar'),
+        'TRY' => array('₺', 'Turkish Lira'),
+        'USD' => array('$', 'United States Dollar'),
+        'ZAR' => array('R', 'South African Rand'),
+    );
+
+    /**
+     * @param string $code
+     * @param string $sign
+     * @param string $name
+     * @param bool $isLeftSign
+     */
+    public static function addCurrency($code, $sign, $name, $isLeftSign = true)
+    {
+        self::$currencies[$code] = array($sign, $name, $isLeftSign);
+    }
+
+    /**
      * @param string $code
      * @return Currency
      * @throws CurrencyNotSupportedException
@@ -58,33 +99,6 @@ class CurrencyFactory
      */
     public static function getAllSupported()
     {
-        // format is:
-        // 'code' => array('sign', 'name')
-        // OR
-        // 'code' => array('sign', 'name', 'isLeftSign')
-        return array(
-            'AUD' => array('$', 'Australian Dollar'),
-            'BRL' => array('R$', 'Brazilian Real'),
-            'CAD' => array('$', 'Canadian Dollar'),
-            'CNY' => array('¥', 'Chinese Yuan'),
-            'CHF' => array('Fr', 'Swiss Franc'),
-            'DKK' => array('kr.', 'Danish Krone'),
-            'EUR' => array('€', 'Euro'),
-            'GBP' => array('£', 'British Pound Sterling'),
-            'HKD' => array('$', 'Hong Kong Dollar'),
-            'INR' => array('₹', 'Indian Rupee'),
-            'ILS' => array('₪', 'Israeli New Shekel'),
-            'JPY' => array('¥', 'Japanese Yen'),
-            'KRW' => array('₩', 'South Korean Won'),
-            'MXN' => array('$', 'Mexican Peso'),
-            'NZD' => array('$', 'New Zealand Dollar'),
-            'NOK' => array('kr', 'Norwegian Krone'),
-            'RUB' => array('₽', 'Russian Ruble'),
-            'SEK' => array('kr', 'Swedish Krona'),
-            'SGD' => array('$', 'Singapore Dollar'),
-            'TRY' => array('₺', 'Turkish Lira'),
-            'USD' => array('$', 'United States Dollar'),
-            'ZAR' => array('R', 'South African Rand'),
-        );
+        return self::$currencies;
     }
 }