+++ changelog.md
@@ -4,6 +4,8 @@
 
 * Switch to PSR-2 style guide
 * Add setDefault method to CurrencyFactory
+* Allow CurrencyConversionService to be passed in to money constructor
+* Add addCurrency method to CurrencyFactory
 
 ## 1.0.2 - 2016-01-27